Remove `ghr` for the builtin `gh` CLI
`ghr` is an old implementation from a time before GitHub had an official
CLI. Now that the CLI exists, is actively maintained, and included in
GitHub Actions by default, it is a great time to switch to the officially
supported way of creating a GitHub release.

This adjusts the `deploy-system` task to use `gh`. It also removes the
need for a special `GITHUB_TOKEN` since the `permissions` key can be
used in a workflow to use builtin action mechanisms to give access needed
for the specific job running. The `README.md` has been updated to reflect
this change as well